[WIP] Bump DEBOUNCE way higher for ErgoDox
I keep getting chatter with my current keymap and board.

Interestingly, I didn't get any before rebasing my keymap on top of
recent master (I was about a year out of date before that).

WIP because changing this for the ErgoDox EZ generally is a bad plan and
because I'm not even sure yet if this is the right number. Just wanted
to see if it would stop the chatter without other ill effects, because
chatter's infuriating.