Move - and = to the top-center keys and map Del and \ directly
I had subconsciously been leaving the top-center untouched, because I
used to use them as the leader key, but I gave up on the leader key a
while ago, so now they're free for other uses.

That in turn frees up a pair of keys, so I can map Del (which I rarely
use but sometimes need for Windows boxes) and \ directly on keys.