[WIP] Port old truly_ergodoxic layout to current qmk
WIP because it's not _quite_ in the state I want.

It's getting closer, but hacking keyboards/ergodox_ez/info.json to
generate .bin files for use with Keymapp does not seem like the optimal
approach to deploying my firmware to my board.

And of course I have yet to actually test the generated output. :P